Fix NULL handling in FQexecParams()
It turns out that if a column has a NOT NULL constraint, but
the provided parameter for that column is NULL, if var->sqldata is
allocated, the column will be filled with whatever the allocated
memory contains (probably garbage, as we weren't initialising it).

This does not happen if the column has no NOT NULL constraint.