Add support for "prepared" queries
This commit adds the following API functions:

 - FQprepare()
 - FQexecPrepared()
 - FQdeallocatePrepared()

These are designed to work in a similar way as the equivalent libpq
functions. However, in contrast to PostgreSQL, a call to FQprepare()
does not result in a named statement being prepared;
instead, a FQresult object is returned with isc_dsql_prepare()
executed. When this is passed to FQexecPrepared(), the query will be
executed in the same way as with FQexecParams(), but the associated
statement handle will not be freed. This substantially reduces
overhead when the same statement is executed multiple times.