Fix _FQinitClientEncoding()
The query used to obtain the client's character set encoding was
assuming MON$REMOTE_PID will be set, however if the client is not
connecting via the server process, it will be NULL.

It turns out CURRENT_CONNECTION returns the value corresponding
to MON$ATTACHMENT_ID, so use that instead.

GitHub #6.