Add support for columns defined with "CHARACTER SET OCTETS"
For CHAR/VARCHAR columns, if the associated XSQLVAR has "sqlsubtype"
set to 1, this means the column was defined with "CHARACTER SET
OCTETS".

As OCTETS are raw binary data, convert these values into hexadecimal
representation, to match what isql does.

It may be desirable to add a connection-level option to return the
raw bytes, but there doesn't seem much point at the moment.