Add support for time zone datatypes
This commit adds support for:

- TIME WITH TIME ZONE
- TIMESTAMP WITH TIME ZONE

(Firebird 4.0 and later).

Note that by default, libfq will emit timestamps with the actual time
zone offset (e.g. "+09:00") rather than the time zone name, if this was
specified in the timestamp.

This behaviour can be changed by setting the (new) connection parameter
"time_zone_names", which will cause the time zone to be emitted with
either the time zone name or the offset.