Improve FLOAT and DOUBLE PRECISION formatting
This brings FLOAT and DOUBLE PRECISION more into line with the output
emitted by clients other than isql.

A new connection option "isql_values" is added to restore the
previous behaviour, if required for any reason.

Note that floating point output is a tricky field of endeavour and
the libfq implementation may contain errors and/or be subject to
change in future releases.